Adding float parameter object

+++ include/audio_generator_render_stage.h
@@ -72,7 +72,7 @@ private:
         };
 
         layout(std140) uniform tone {
-            int tone_val;
+            float tone_val;
         };
 
         out vec4 output_audio_texture;
@@ -80,7 +80,7 @@ private:
         vec2 translate_coord(vec2 coord) {
             // Get the chunk size
             ivec2 audio_size = textureSize(full_audio_data_texture, 0);
-            ivec2 chunk_size = ivec2(1024.0 * float(tone_val), 1);
+            ivec2 chunk_size = ivec2(1024.0 * tone_val, 1);
 
             int chunk_offset = time_val * chunk_size.x;
